# grammars: early exit when no next_candidates to reject